# api/openapi.yaml
openapi: 3.0.4
info:
  title: ADB Console API
  version: 1.0.0
  description: This is an API documentation for ADB console server.
servers:
  - url: http://localhost:8080
    description: default
paths:
  /gui:
    get:
      summary: Get GUI
      description: Get main page of ADB console.
      operationId: GetMainPage
      responses:
        '200':
          description: Return main page of ADB console.
          content:
            text/html:
              schema:
                type: string
  /api/devices:
    get:
      summary: Execute adb devices command and report device status.
      description: DeviceList is used for data within SEE response.
      tags:
        - general command
      operationId: GetDevices
      parameters:
        - $ref: "#/components/parameters/PollingInterval"
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/adb/version:
    get:
      summary: Get version of application, ADB and SDK.
      tags:
        - general command
      operationId: GetVersion
      responses:
        '200':
          description: Retusn version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Version"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails" 
  /api/adb/push:
    post:
      summary: Execute adb push command that push file(s) and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - device command
      operationId: ExecuteAdbPush
      parameters:
        - name: args
          description: Command arguments.
          in: query
          required: true
          style: form
          explode: false
          schema:
            $ref: "#/components/schemas/CommandArgs"
          example: args=dosc%2Fsample.png%2C%2Fsdcard%2Fpicture%2F
      requestBody:
        $ref: "#/components/requestBodies/TargetList"
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails" 
  /api/adb/pull:
    post:
      summary: Execute adb pull command that pull file(s) and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - device command
      operationId: ExecuteAdbPull
      parameters:
        - name: args
          description: Command arguments.
          in: query
          required: true
          style: form
          explode: false
          schema:
            $ref: "#/components/schemas/CommandArgs"
          example: args=%2Fsdcard%2Fpicture%sample.png
      requestBody:
        $ref: "#/components/requestBodies/TargetList"
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"    
  /api/adb/shell:
    post:
      summary: Execute adb shell command and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - device command
      operationId: ExecuteAdbShell
      parameters:
        - name: args
          description: Command arguments.
          in: query
          required: true
          style: form
          explode: false
          schema:
            $ref: "#/components/schemas/CommandArgs"
          example: args=ping%2C-n%2C4%2C8.8.8.8
      requestBody:
        $ref: "#/components/requestBodies/TargetList"
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"            
  /api/adb/logcat:
    post:
      summary: Execute adb logcat command and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - device command
      operationId: ExecuteAdbLogcat
      parameters:
        - name: args
          description: Command arguments.
          in: query
          required: false
          style: form
          explode: false
          schema:
            $ref: "#/components/schemas/CommandArgs"
          example: args=I%2Czone
      requestBody:
        $ref: "#/components/requestBodies/TargetList"
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"            

  /api/adb/reboot:
    post:
      summary: Execute adb reroot command and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - device command
      operationId: ExecuteAdbReboot
      parameters:
        - name: args
          description: Command arguments.
          in: query
          required: false
          style: form
          explode: false
          schema:
            $ref: "#/components/schemas/RebootArgs"
          example: bootloader
      requestBody:
        $ref: "#/components/requestBodies/TargetList"
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"  
  /api/adb/root:
    post:
      summary: Execute adb root command and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - device command
      operationId: ExecuteAdbRoot
      requestBody:
        $ref: "#/components/requestBodies/TargetList"
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"  
  /api/adb/unroot:
    post:
      summary: Execute adb unroot command and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - device command
      operationId: ExecuteAdbUnroot
      requestBody:
        $ref: "#/components/requestBodies/TargetList"
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/adb/start-server:
    post:
      summary: Execute adb start-server command and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - general command
      operationId: ExecuteAdbStartServer
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/adb/kill-server:
    post:
      summary: Execute adb kill-server command and report result.
      description: CommandResults is used for data within SSE response.
      tags:
        - general command
      operationId: ExecuteAdbKillServer
      responses:
        '200':
          $ref: "#/components/responses/SSEResponse"
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

components:
  parameters:
    CommandArgs:
      description: List of command arguments.
      name: args
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: "#/components/schemas/CommandArgs"
      example: args=-n,4,8.8.8.8
    PollingInterval:
      description: Interval time for polling. (seconds) If omit or 0 then no polling. (1 shot)
      name: interval
      in: query
      required: false
      style: form
      explode: true
      schema:
        type: integer
        minimum: 0
      example: 5
  requestBodies:
    TargetList:
      description: List of command target.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SerialList"
      required: true
  responses:
    SSEResponse:
      description: Response SSE stream
      headers:
        Content-Type:
          schema:
            description: Content-Type for SSE
            type: string
            example: text/event-stream
        Cache-Control:
          schema:
            description: Prohibit cashe in browser and/or relay server.
            type: string
            example: no-cache
        Connection:
          schema:
            description: Keep connection.
            type: string
            example: keep-alive
        X-Accel-Buffering:
          schema:
            description: Disable response buffering in proxy. (option)
            type: string
            example: 'no'
      content:
        text/event-stream:
          schema:
            description: SSE message. (data body is JSON format)
            type: string
            example: 'data: {...}'

  schemas:
    CommandArgs:
      description: Command arguments.
      type: array
      items:
        type: string
    RebootArgs:
      description: argument list for adb reboot.
      type: string
      enum:
        - bootloader
        - recovery
        - sideload
        - sideload-auto-reboot
    CommandResult:
      description: Command result with serial.
      type: object
      properties:
        serial:
          $ref: "#/components/schemas/DeviceSerial"
        result:
          type: string
          example: "PING 8.8.8.8 (8.8.8.8) 56(84) bytes of data."
      required:
        - serial
        - result
    CommandResults:
      description: List of command results.
      type: array
      items:
        type: object
        $ref: "#/components/schemas/CommandResult"
    DeviceSerial:
      description: serial of device.
      type: string
      example: ABC123DEF
    DeviceStatus:
      description: Status of device.
      type: string
      enum:
        - online
        - offline
        - unauthorized
        - unknown
      example: online
    DeviceInfo:
      description: Devide information.
      type: object
      properties:
        serial:
          $ref: "#/components/schemas/DeviceSerial"
        status:
          $ref: "#/components/schemas/DeviceStatus"
        product:
          description: Product name of device.
          type: string
          example: ???
        model:
          description: Model name of device.
          type: string
          example: Pixel-10
        device:
          type: string
          example: ???
        tid:
          description: Transport ID.
          type: integer
          example: 1
      required:
        - serial
        - status
    DeviceList:
      description: List of devices with information.
      type: array
      items:
        $ref: "#/components/schemas/DeviceInfo"
    SerialList:
      description: List of device serial.
      type: array
      items:
        $ref: "#/components/schemas/DeviceSerial"
      minLength: 1
    Version:
      description: Version information
      properties:
        app:
          description: Application version.
          type: string
          example: 1.0.0
        api:
          description: API version.
          type: string
          example: 1.0.0
        adb:
          description: ADB version
          type: string
          example: 1.0.0
        sdk:
          description: Android SDK version
          type: string
          example: 1.0.0
      required:
        - app
    ProblemDetails:
      description: Details for error response.
      type: object
      properties:
        status:
          description: Status code.
          type: integer
          example: 500
        title:
          description: Readful text for status code.
          type: string
          example: Internal Server Error
        detail:
          description: Error message.
          type: string
          example: "'foo' is not recognized as an internal or external command, operable program or batch file."

tags:
  - name: general command
    description: General command.
  - name: device command
    description: Command for each device. (i.e. need serial)